+++ application/views/base/players.php
@@ -1,0 +1,19 @@
+<h2 class="text-center underline">Players</h2>
+<table class="table table-striped">
+    <thead>
+        <tr>
+            <th>Name</th>
+            <th>Cash</th>
+            <th>Equity</th>
+        </tr>
+    </thead>
+    <tbody>
+        {players}
+        <tr>
+            <td><a href={href}>{Player}</a></td>
+            <td>{Cash}</td>
+            <td>{Equity}</td
+        </tr>
+        {/players}
+    </tbody>
+</table>