+++ application/views/player/players.php
@@ -1,0 +1,7 @@
+<h2 id="playername" class="text-center underline">{Playername}</h2>
+{form}
+{select}
+
+<div id="holdings-chart"></div>
+
+<script src="../assets/js/holding-history.js"></script>