+++ application/views/stock/index.php
@@ -1,0 +1,12 @@
+<h1 class="text-center">
+    {Name} ({Code}): &nbsp; {Value}, {Category}
+</h1>
+
+{form}
+{select}
+
+<div id="stock_code" class="hidden-xs hidden-sm hidden-md hidden-lg">{stock_code}</div>
+<div id="stock_code" class="hidden-xs hidden-sm hidden-md hidden-lg">{url}</div>
+
+<div id="temp-stocks"></div>
+<script src={src}></script>