+++ application/views/templates/_logout_control.php
@@ -1,0 +1,4 @@
+<form class='navbar-form navbar-right' action='/user/logoff' method='post'>
+     Hello, {username}
+    <button type='submit' class='btn btn-danger'>Logoff</button>
+</form>